Assume the user configures passenger
It doesn't really matter how its setup as long as its funcitonal so it doesn't
make sense to include the setup as part of this recipe. Just let the user
handle it and give them more flexibility.

This is how it was mostly setup anyway its just more clearly documented now.